requirements.txt
Pillow
opencv-python
PyPDF2
easygui
cx_Freeze